requirements.txt
vk_api==11.9.1
python-dotenv==0.18.0